Add create trip transaction when Keer feedback

+++ Application/Feedbacks/FeedbackDTO.cs
@@ -5,7 +5,7 @@ namespace Application.Feedbacks.DTOs
 		public int UserId { get; set; }
 		public int TripId { get; set; }
 		public string FeedbackContent { get; set; }
-		public double TripStar { get; set; }
+		public int TripStar { get; set; }
 		public string Criteria { get; set; }
 	}
 }
